refactor: 공통 Emptystate 사용으로 변경
